vim: fix spelling highlights for terminal
The default highlighting of spelling errors is calibrated to make people
want to poke their eyes out with a swizzle stick. Since source code
always has some technical gibberish in the comments, I need to make it
something more tolerable to be able to use the terminal vim for code.